Fix detection of open rings.
When the ring was open and one of the ring ports was going down. After
it came back up the ring state was ring closed. It never went back to
open ring even if kernel was notifying the userspace that the ring is
open.

The issue was that the ring ports had the flags loc('lost of
continuity') and these flags were not reset when ring state was
changed to ring closed. So then when the kernel notified that the ring
is open, the userspace ignored these notifications because already knew
that the ring is open, so no action was taken.

The fix consist in clearing loc flag when the ring goes in closed ring.