Fix for MRA transition from MRC to MRM
The following case was not working.
Node 1 which is MRA is started and gets the role of MRM and because the
ring is open it gets the CHK_RO state.
Node 2 which is MRA is started and gets the role of MRM because it has a
better priority and it goes in CHC_RC and Node 1 goes in PT_IDLE.
Now if Node 2 is removed, then Node 1 didn't managed to change from MRC
to MRM even if the kernel notified it that it stopped seeing that there
are no Test frames.

The reason was that the Node 1 already process the lost of continuity
when it went in CHK_RO and when it receives another one it just ignored it.

The fix consists in resetting the lost of continuity when it changes its
role from MRM to MRC.