Fix for passing adding cfm_maid and cfm_dmac
When using MRP interconnect in LC mode, then it is required to use CFM
frames. In this case it is required to add the cfm mac address and cfm
maid. The problem was that these arguments were copy using strncpy and
if they had any null characters then it stopped to copy the rest. So for
example if dmac was 00:11:22:33:44:55 then it copy was set to be
00:00:00:00:00:00. This commit fixes this issue by using memcpy instead
strncpy.