Pass to the kernel the best MAC address
When the role is MRA, it is required to compare the priority in the
frame with the priority of the node and also to compare the MAC address
of the frame(MRP_SA) with the best known MAC in the ring.
To be able to do this comparison, it is required to push the best known
MAC address to the kernel because the kernel compares with it's own MAC
address.