Initialize ring_mac to bridge mac
The ring_mac is initialized to zero and changed based on received
frames. If the ring_mac is used to dispatch frames, before it is set to
a non-zero value the link-partners will receive a zero best_mac, which
will update their best mac as it is lower than anything.

Avoid this by initializing ring_mac to the bridge mac.